+++ src/stdb/battle_settings.rs
@@ -16,6 +16,6 @@ use spacetimedb_sdk::{
 pub struct BattleSettings {
     pub fatigue_start: u32,
     pub deafness_start: u32,
-    pub deafness_per_turn: f64,
+    pub deafness_per_turn: f32,
     pub summon_limit: u32,
 }
